profile-chat-server/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
pypdf
mangum
boto3
